misc minor JAI fixes

+++ src/JSystem/JAudio/JAIBasic.cpp
@@ -297,7 +297,7 @@ f32 JAIBasic::getMapInfoFxParameter(u32 param_1) {
 }
 
 /* 80290DC4-80290E14       .text getSoundOffsetNumberFromID__8JAIBasicFUl */
-int JAIBasic::getSoundOffsetNumberFromID(u32) {
+u16 JAIBasic::getSoundOffsetNumberFromID(u32) {
     /* Nonmatching */
 }
